@@ -11,42 +11,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
+    "@angular/animations": "~10.0.0-next.7",
     "@angular/cdk": "~8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
+    "@angular/common": "~10.0.0-next.7",
+    "@angular/compiler": "~10.0.0-next.7",
+    "@angular/core": "~10.0.0-next.7",
     "@angular/flex-layout": "^9.0.0-beta.31",
-    "@angular/forms": "~8.2.14",
+    "@angular/forms": "~10.0.0-next.7",
     "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "rxjs": "~6.4.0",
+    "@angular/platform-browser": "~10.0.0-next.7",
+    "@angular/platform-browser-dynamic": "~10.0.0-next.7",
+    "@angular/router": "~10.0.0-next.7",
+    "rxjs": "~6.5.5",
     "tslib": "^1.10.0",
     "yarn": "^1.22.4",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.21",
-    "@angular/cli": "~8.3.21",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
+    "@angular-devkit/build-angular": "~0.1000.0-next.5",
+    "@angular/cli": "~10.0.0-next.5",
+    "@angular/compiler-cli": "~10.0.0-next.7",
+    "@angular/language-service": "~10.0.0-next.7",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@types/node": "^12.11.1",
     "@types/rx": "^4.1.1",
-    "codelyzer": "^5.0.0",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
+    "protractor": "~5.4.4",
     "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "tslint": "~6.1.0",
+    "typescript": "^3.8.3"
   }
 }
